polish + add getAllProjectGroups
