+++ src/screens/general/AllBills.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../../styles/AllBills.css';
-import { centsToDollars } from '../../lib/subscriberHelper';
+import { centsToDollars } from '../../lib/subscriberUtils';
 import Bill from '../../components/Bill';
 import { getLoggedInUserId, logOut } from '../../lib/auth';
