+++ src/screens/general/SubscriberOwnerDashboard/SubscriberOwnerDashboard.js
@@ -5,10 +5,7 @@ import SubscriberOwnerDashboardAllBillsView from './SubscriberOwnerDashboardAllB
 import SubscriberOwnerDashboardMainView from './SubscriberOwnerDashboardMainView';
 import { getLoggedInUserId } from '../../../lib/auth';
 
-import {
-  areDiffBills,
-  getSubscriberBills
-} from '../../../lib/subscriberHelper';
+import { areDiffBills, getSubscriberBills } from '../../../lib/subscriberUtils';
 
 export default class SubscriberOwnerDashboard extends React.Component {
   constructor(props) {
